+++ src/main/java/com/concrete/concrete_plant_management/order_batch/OrderBatchWriteModel.java
@@ -1,6 +1,6 @@
 package com.concrete.concrete_plant_management.order_batch;
 
-import com.concrete.concrete_plant_management.order.Order;
+import com.concrete.concrete_plant_management.domain.Order;
 import com.concrete.concrete_plant_management.domain.Vehicle;
 import lombok.Getter;
 import lombok.Setter;
@@ -10,8 +10,8 @@ import java.time.LocalTime;
 @Getter
 class OrderBatchWriteModel {
     @Setter
-    private int id;
-    private int batchId;
+    private Long id;
+    private Long batchId;
     private double amount;
     private LocalTime time;
     private boolean toDelete;
